Separate the mesh quality testing from mesh handling
